+++ app/models/slickr/media_upload.rb
@@ -2,13 +2,132 @@ require 'image_processing/mini_magick'
 
 module Slickr
   class MediaUpload < ApplicationRecord
+    self.table_name = 'slickr_media_uploads'
     DROP_AREA_TEXT = 'Maximum size 10Mb | .jpeg, .jpg, .png and .pdf files only'
+
+    ADDITIONAL_DERIVATIVES = {}
+
+    DEFAULT_IMAGE_DERIVATIVES = {
+      square: {
+        square_2400: {
+          type: :fill, options: [2400, 2400, { crop: :attention }]
+        },
+        square_1800: {
+          type: :fill, options: [1800, 1800, { crop: :attention }]
+        },
+        square_1200: {
+          type: :fill, options: [1200, 1200, { crop: :attention }]
+        },
+        square_800: {
+          type: :fill, options: [800, 800, { crop: :attention }]
+        },
+        square_600: {
+          type: :fill, options: [600, 600, { crop: :attention }]
+        },
+        square_400: {
+          type: :fill, options: [400, 400, { crop: :attention }]
+        }
+      },
+      landscape: {
+        landscape_2400: {
+          type: :fill, options: [2400, 1600, { crop: :attention }]
+        },
+        landscape_1800: {
+          type: :fill, options: [1800, 1200, { crop: :attention }]
+        },
+        landscape_1200: {
+          type: :fill, options: [1200, 800, { crop: :attention }]
+        },
+        landscape_800: {
+          type: :fill, options: [800, 534, { crop: :attention }]
+        },
+        landscape_600: {
+          type: :fill, options: [600, 400, { crop: :attention }]
+        },
+        landscape_400: {
+          type: :fill, options: [400, 267, { crop: :attention }]
+        }
+      },
+      portrait: {
+        portrait_2400: {
+          type: :fill, options: [1600, 2400, { crop: :attention }]
+        },
+        portrait_1800: {
+          type: :fill, options: [1200, 1800, { crop: :attention }]
+        },
+        portrait_1200: {
+          type: :fill, options: [800, 1200, { crop: :attention }]
+        },
+        portrait_800: {
+          type: :fill, options: [534, 800, { crop: :attention }]
+        },
+        portrait_600: {
+          type: :fill, options: [400, 600, { crop: :attention }]
+        },
+        portrait_400: {
+          type: :fill, options: [267, 400, { crop: :attention }]
+        }
+      },
+      panoramic: {
+        panoramic_2400: {
+          type: :fill, options: [2400, 1350, { crop: :attention }]
+        },
+        panoramic_1800: {
+          type: :fill, options: [1800, 1012, { crop: :attention }]
+        },
+        panoramic_1200: {
+          type: :fill, options: [1200, 675, { crop: :attention }]
+        },
+        panoramic_800: {
+          type: :fill, options: [800, 450, { crop: :attention }]
+        },
+        panoramic_600: {
+          type: :fill, options: [600, 337, { crop: :attention }]
+        },
+        panoramic_400: {
+          type: :fill, options: [400, 225, { crop: :attention }]
+        }
+      },
+      content: {
+        content_1200: {
+          type: :limit, options: [1200, nil]
+        },
+        content_800: {
+          type: :limit, options: [800, nil]
+        },
+        content_600: {
+          type: :limit, options: [600, nil]
+        },
+        content_400: {
+          type: :limit, options: [400, nil]
+        }
+      }
+    }
+
+    include AASM
+    aasm do
+      state :uploaded, initial: true
+      state :processing
+      state :resizing
+      state :ready
+
+      event :process do
+        transitions from: :uploaded, to: :processing
+      end
+      event :resize, after: :send_for_resizing do
+        transitions from: :processing, to: :resizing
+      end
+      event :finalize do
+        transitions from: :resizing, to: :ready
+      end
+    end
+
+    after_create :send_to_processing
 
     if defined?(acts_as_taggable_on)
       acts_as_taggable_on :media_tags
     end
 
-    self.table_name = 'slickr_media_uploads'
 
     include Slickr::MediaImageUploader[:image] rescue NameError
     include Slickr::MediaFileUploader[:file] rescue NameError
@@ -114,9 +233,9 @@ module Slickr
         id: id,
         src: image_url(:xl_limit),
         displayPath: image_url(:m_limit),
-        thumbnail: image_url(:s_limit),
-        thumbnailWidth: image(:s_limit).width,
-        thumbnailHeight: image(:s_limit).height,
+        thumbnail: image_url(:thumb_400x400),
+        thumbnailWidth: image(:thumb_400x400).width,
+        thumbnailHeight: image(:thumb_400x400).height,
         caption: image.original_filename,
         isSelected: false,
         editPath: admin_edit_path,
@@ -133,6 +252,22 @@ module Slickr
         thumbnailWidth: 127, thumbnailHeight: 180,
         caption: '', isSelected: false, editPath: '', mimeType: ''
       }
+    end
+
+    def send_to_processing
+      process
+    end
+
+    def send_for_resizing
+      ResizeImagesJob.perform_later(self, :square)
+      ResizeImagesJob.perform_later(self, :portrait)
+      ResizeImagesJob.perform_later(self, :landscape)
+      ResizeImagesJob.perform_later(self, :panoramic)
+      ResizeImagesJob.perform_later(self, :content)
+    end
+
+    def add_new_derivatives(key)
+      ResizeImagesJob.perform_later(self, key)
     end
   end
 end