+++ db/migrate/20171129144524_add_header_image_to_slickr_pages.rb
@@ -1,5 +1,0 @@
-class AddHeaderImageToSlickrPages < ActiveRecord::Migration[5.1]
-  def change
-    add_column :slickr_pages, :page_header_image, :string
-  end
-end