+++ lib/generators/slickr/install_generator.rb
@@ -248,17 +248,17 @@ module Slickr
         puts "Slickr Page extended"
       end
 
-      def slickr_media_image_uploader
-        template "media_image_uploader.rb", "app/uploaders/slickr/media_image_uploader.rb"
-
-        puts "Slickr::MediaImageUploader added"
-      end
-
-      def slickr_media_file_uploader
-        template "media_file_uploader.rb", "app/uploaders/slickr/media_file_uploader.rb"
-
-        puts "Slickr::MediaFileUploader added"
-      end
+      # def slickr_media_image_uploader
+      #   template "media_image_uploader.rb", "app/uploaders/slickr/media_image_uploader.rb"
+      #
+      #   puts "Slickr::MediaImageUploader added"
+      # end
+      #
+      # def slickr_media_file_uploader
+      #   template "media_file_uploader.rb", "app/uploaders/slickr/media_file_uploader.rb"
+      #
+      #   puts "Slickr::MediaFileUploader added"
+      # end
 
       def extend_active_admin_initializer
         dest_file = "config/initializers/active_admin.rb"